Implement GSuiteMigration class
This commit adds a class that will generate `terraform state mv`
commands to update a Terraform statefile to match the changes made in
the G Suite refactor. There are some outstanding questions about how to
handle Terraform configurations that use the G Suite resources since
it's non-trivial to determine when this change happened, so this commit
will need follow up work to fully implement it.